GIS refactoring and using d3-delaunay.
