+++ package.json
@@ -20,6 +20,7 @@
     "commander": "^2.14.1",
     "console.table": "0.10.0",
     "crypto-js": "3.1.9-1",
+    "d3-delaunay": "5.1.6",
     "event-stream": "^4.0.1",
     "fast-csv": "3.4.0",
     "geojson-validation": "^0.2.1",
@@ -60,7 +61,8 @@
     "rewire": "^4.0.1",
     "ts-node": "^7.0.1",
     "typescript": "^3.1.1",
-    "@types/table": "^4.0.6"
+    "@types/table": "^4.0.6",
+    "@types/d3-delaunay": "4.1.0"
   },
   "bin": {
     "here": "bin/here.js"
