+++ src/here-xyz.ts
@@ -31,10 +31,10 @@ import * as common from "./common";
 import * as sso from "./sso";
 import * as inquirer from "inquirer";
 import * as transform from "./transformutil";
+import * as gis from "./gisUtil";
 import * as fs from "fs";
 import * as tmp from "tmp";
 import * as summary from "./summary";
-import * as turf from "@turf/turf";
 let cq = require("block-queue");
 import { isBoolean } from "util";
 import { ApiError } from "./api-error";
@@ -367,7 +367,7 @@ program
         })();
     });
 
-function getSpaceDataFromXyz(id: string, options: any) {
+export function getSpaceDataFromXyz(id: string, options: any) {
     return new Promise<any>(function (resolve, reject) {
         let cType = "application/json";
         if (!options.limit) {
@@ -772,7 +772,7 @@ async function isOtherOwnerSpace(spaceOwner: string) {
     return currentOwner != spaceOwner;
 }
 
-async function createNewSpaceUpdateMetadata(newSpaceType: string, sourceId: string, sourceSpaceData: any, updateSourceMetadata: boolean = true, newSpacetoken: string | null = null) {
+export async function createNewSpaceUpdateMetadata(newSpaceType: string, sourceId: string, sourceSpaceData: any, updateSourceMetadata: boolean = true, newSpacetoken: string | null = null) {
     let newSpaceConfig = {
         title: newSpaceType+ ' space of ' + sourceSpaceData.title,
         message: newSpaceType + ' space created for source spaceId - ' + sourceSpaceData.id + ' , title - ' + sourceSpaceData.title,
@@ -821,7 +821,7 @@ async function getBoundingBoxFromUser() {
     return answer.minx + "," + answer.miny + "," + answer.maxx + "," + answer.maxy;
 }
 
-async function getSpaceMetaData(id: string, token: string | null = null) {
+export async function getSpaceMetaData(id: string, token: string | null = null) {
     const uri = "/hub/spaces/" + id + "?clientId=cli";
     const cType = "application/json";
     const { response, body } = await execute(uri, "GET", cType, "", token);
@@ -1345,7 +1345,7 @@ function taskQueue(size: number = 8, totalTaskSize: number) {
 }
 
 
-async function uploadToXyzSpace(id: string, options: any) {
+export async function uploadToXyzSpace(id: string, options: any) {
     let startTime = new Date();
     //(async () => {
     let tags = "";
@@ -2337,7 +2337,7 @@ function getEmptyAcitivityLogListenerProfile() {
 }
 
 
-async function getSpaceStatistics(id: string, token: string | null = null) {
+export async function getSpaceStatistics(id: string, token: string | null = null) {
     const uri = "/hub/spaces/" + id + "/statistics?clientId=cli";
     const cType = "application/json";
     const { response, body } = await execute(uri, "GET", cType, "", token);
@@ -3042,177 +3042,12 @@ program
     .option("--property <property>", "populates tin polygons' properties based on the feature property specified here")
     .option("--samespace", "option to upload centroids/voronoi/tin to same space")
     .action(function (id, options) {
-        performGisOperation(id, options).catch((error) => {
+        gis.performGisOperation(id, options).catch((error) => {
             handleError(error, true);
         });
     });
 
-async function performGisOperation(id:string, options:any){
-    await common.verifyProBetaLicense();
-    const sourceId = id;
-    options.totalRecords = Number.MAX_SAFE_INTEGER;
-    options.currentHandleOnly = true;
-    options.handle = 0;
-    if(options.chunk){
-        options.limit = options.chunk;
-    } else {
-        options.limit = 20;
-    }
-    if(!options['length'] && !options.centroid && !options.area && !options.voronoi && !options.tin){
-        console.log("Please specify GIS operation option");
-        process.exit(1);
-    }
-    let cHandle;
-    let gisFeatures : any[]= [];
-    let featureCount = 0;
-    console.log("Performing GIS operation on the space data");
-    do {
-        let jsonOut = await getSpaceDataFromXyz(id, options);
-        if (jsonOut.features && jsonOut.features.length === 0 && options.handle == 0) {
-            console.log("\nNo features are available to execute GIS operation");
-            process.exit(1);
-        }
-        cHandle = jsonOut.handle;
-        options.handle = jsonOut.handle;
-        if (jsonOut.features) {
-            const features = jsonOut.features;
-            features.forEach(function (feature: any, i: number){
-                if(options.voronoi || options.tin){
-                    if(feature.geometry && (feature.geometry.type == 'Point')){
-                        gisFeatures.push(feature);
-                    }
-                } else {
-                    let gisFeature = performTurfOperationOnFeature(feature, options);
-                    if(gisFeature){
-                        gisFeatures.push(gisFeature);
-                        process.stdout.write("\rGIS operation done for feature count - " + (featureCount + (i+1)));
-                    }
-                }
-            });
-            featureCount += features.length;
-        } else {
-            cHandle = -1;
-        }
-    } while (cHandle >= 0);
-    process.stdout.write("\n");
-
-    if(gisFeatures.length == 0){
-        console.log("required geometry features are not available to perform GIS operation");
-        process.exit(1);
-    }
-
-    if(options.voronoi){
-        console.log("Calculating Voronoi Polygons for points data");
-        gisFeatures = await calculateVoronoiPolygon(id, gisFeatures, options);
-    } else if(options.tin){
-        gisFeatures = calculateTinTriangles(gisFeatures, options.property);
-    }
-    if(!options.samespace && (options.centroid || options.voronoi || options.tin)){
-        let newSpaceData;
-        if(options.centroid){
-            newSpaceData = await createNewSpaceAndUpdateMetadata('centroid', sourceId, options);
-        } else if(options.voronoi){
-            newSpaceData = await createNewSpaceAndUpdateMetadata('voronoi', sourceId, options);
-        } else if(options.tin){
-            newSpaceData = await createNewSpaceAndUpdateMetadata('tin', sourceId, options);
-        }
-        id = newSpaceData.id;
-    }
-    
-    if (gisFeatures.length > 0) {
-        let tmpObj = tmp.fileSync({ mode: 0o644, prefix: 'gis', postfix: '.json' });
-        fs.writeFileSync(tmpObj.name, JSON.stringify({ type: "FeatureCollection", features: gisFeatures }));
-        if(options.centroid){
-            options.tags = 'centroid';
-        } else if(options.voronoi){
-            options.tags = 'voronoi';
-        } else if(options.tin){
-            options.tags = 'tin';
-        }
-        options.file = tmpObj.name;
-        options.override = true;
-        await uploadToXyzSpace(id, options);
-        console.log("GIS operation completed on space " + sourceId);
-    }
-}
-
-function performTurfOperationOnFeature(feature: any, options: any){
-    let gisFeature;
-    if(options.centroid){
-        if(feature.geometry && (feature.geometry.type == 'LineString' || feature.geometry.type == 'Polygon' || feature.geometry.type == 'MultiLineString' || feature.geometry.type == 'MultiPolygon')){
-            gisFeature = turf.centroid(feature, feature.properties);
-            if(options.samespace){
-                if(!gisFeature.properties){
-                    gisFeature.properties = {};
-                }
-                gisFeature.properties.sourceId = feature.id;
-            } else {
-                gisFeature.id = feature.id;
-            }
-        }
-    } else if(options['length']){
-        if(feature.geometry && (feature.geometry.type == 'LineString' || feature.geometry.type == 'MultiLineString')){
-            let length = turf.length(feature, {units: 'meters'});
-            if(!feature.properties){
-                feature.properties = {};
-            }
-            feature.properties['xyz_length_m'] = length;
-            feature.properties['xyz_length_km'] = (length / 1000).toFixed(2);
-            feature.properties['xyz_length_miles'] = (length * 0.000621371).toFixed(2);
-            gisFeature = feature; 
-        }
-    } else if(options.area){
-        if(feature.geometry && (feature.geometry.type == 'Polygon' || feature.geometry.type == 'MultiPolygon')){
-            gisFeature = populateArea(feature); 
-        }
-    } else {
-        console.log("Please specify GIS operation option");
-        process.exit(1);
-    }
-    return gisFeature;
-}
-
-function populateArea(feature: any){
-    let area = turf.area(feature);
-    if(!feature.properties){
-        feature.properties = {};
-    }
-    feature.properties['xyz_area_sqm'] = area;
-    feature.properties['xyz_area_sqkm'] = (area / 1000000).toFixed(2);
-    feature.properties['xyz_area_sqmiles'] = (area * 0.00000038610215855).toFixed(2);
-    return feature; 
-}
-
-async function calculateVoronoiPolygon(spaceId: string, features: any[], options: any){
-    const statData = await getSpaceStatistics(spaceId);
-    const featureCollection = { type: "FeatureCollection", features: features };
-    const bbox: [number,number,number, number] = statData.bbox.value;
-    const voronoiFeatureCollection = turf.voronoi(featureCollection, {bbox: bbox});
-    voronoiFeatureCollection.features.forEach(function (polygon, i) {
-        polygon.properties = features[i].properties;
-        if(options.samespace){
-            if(!polygon.properties){
-                polygon.properties = {};
-            }
-            polygon.properties.sourceId = features[i].id;
-        } else {
-            polygon.id = features[i].id;
-        }
-        polygon = populateArea(polygon);
-    });
-    return voronoiFeatureCollection.features;
-}
-
-function calculateTinTriangles(features: any, property: string){
-    const featureCollection = { type: "FeatureCollection", features: features };
-    const tinFeatureCollection = turf.tin(featureCollection, property);
-    tinFeatureCollection.features.forEach(function (polygon, i) {
-        polygon = populateArea(polygon);
-    });
-    return tinFeatureCollection.features;
-}
-
-async function createNewSpaceAndUpdateMetadata(newSpaceType: string, sourceId: string, options: any){
+export async function createNewSpaceAndUpdateMetadata(newSpaceType: string, sourceId: string, options: any){
     let newSpaceId;
     let sourceSpaceData = await getSpaceMetaData(sourceId, options.readToken);
     let newspaceData;
